requirements.txt
djangorestframework==3.9.0
psycopg2==2.7.5
dj-database-url==0.5.0
django-filter==2.0.0